% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20241206_simulation_function_edit.R
\name{K}
\alias{K}
\title{Uniform kernel function}
\usage{
K(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a vector of values
}
\description{
Uniform kernel function
}
\examples{
K(0)
}
