% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20241206_simulation_function_edit.R
\name{MAinf_normal}
\alias{MAinf_normal}
\title{Simulate an MA infinity process with algrebraic decay}
\usage{
MAinf_normal(N, beta)
}
\arguments{
\item{N}{sample size}

\item{beta}{algebraic decay parameter}
}
\value{
simulated MA infinity process
}
\description{
Simulate an MA infinity process with algrebraic decay
}
\examples{
x = MAinf_normal(100, 1.5)
}
