% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20241206_simulation_function_edit.R
\name{threshold.example}
\alias{threshold.example}
\title{Simulate an example data frame}
\usage{
threshold.example(N, TL, p, gamma)
}
\arguments{
\item{N}{cross-sectional dimension}

\item{TL}{time series length}

\item{p}{fraction of non-zero coefficients}

\item{gamma}{value of non-zero coefficients}
}
\value{
simulated data frame
}
\description{
Simulate an example data frame
}
\examples{
d = threshold.example(10, 200, 0.1, 2)
}
